Rename util package to core
The package should contain classes that need to be accessible to all
projects/modules but which are not intended for benchmark writers or
suite users. This includes both the Renaissance harness and the JMH
benchmark wrappers.

While "harness" would be also a suitable name for the package, we
want to avoid using the same package name for the core classes and
the harness classes, because they would still be considered two
different packages (core and harness classes are loaded by different
class loaders).

This arrangement ensures that IDEs understand that "core" classes used
directly by the harness are publicly accessible, even though developers
might be tempted to make some of the package private.
