+++ renaissance-core/src/main/java/org/renaissance/core/BenchmarkInfo.java
@@ -1,4 +1,4 @@
-package org.renaissance.util;
+package org.renaissance.core;
 
 import java.util.Map;
 import java.util.NoSuchElementException;
