+++ renaissance-core/src/main/java/org/renaissance/core/BenchmarkRegistry.java
@@ -1,4 +1,4 @@
-package org.renaissance.util;
+package org.renaissance.core;
 
 import org.renaissance.Benchmark;
