+++ renaissance-core/src/main/java/org/renaissance/core/DirUtils.java
@@ -1,4 +1,4 @@
-package org.renaissance.util;
+package org.renaissance.core;
 
 import java.io.IOException;
 import java.nio.file.*;
