+++ renaissance-harness/src/main/java/org/renaissance/ExecutionPolicyFactory.java
@@ -1,7 +1,7 @@
 package org.renaissance;
 
 import org.renaissance.harness.ExecutionPolicy;
-import org.renaissance.util.BenchmarkInfo;
+import org.renaissance.core.BenchmarkInfo;
 
 import static java.util.concurrent.TimeUnit.SECONDS;
